% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel.R
\name{closeTunnel}
\alias{closeTunnel}
\title{Close an SSH tunnel}
\usage{
closeTunnel(con)
}
\arguments{
\item{con}{An object of class "tunnel" returned by \code{tunnel()}.}
}
\value{
\code{TRUE} if the function closes the tunnel associated
  with \code{con}, \code{NA} if there was no tunnel to close, or
  \code{FALSE} if the tunnel could not be closed.
}
\description{
Close an SSH tunnel
}
\examples{
\dontrun{
sshTunnel = tunnel(remote_server, server_username,
                   remote_port, local_port)
closeTunnel(sshTunnel)
}
}

