% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel.R
\name{tunnel}
\alias{tunnel}
\title{Set up SSH tunneling}
\usage{
tunnel(remote_server, server_username, remote_port = 3306,
  local_port = 9000)
}
\arguments{
\item{remote_server}{The URL or IP address of the remote server.}

\item{server_username}{Username to log into the remote server.}

\item{remote_port}{The port to connect to on the remote
server. 3306 is the default port used by MySQL.}

\item{local_port}{The port to connect to on the local machine.}
}
\value{
A connection object of class "tunnel".
}
\description{
Set up SSH tunneling
}
\examples{
\dontrun{
sshTunnel = tunnel(remote_server, server_username,
                   remote_port, local_port)
closeTunnel(sshTunnel)
}
}

