% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel.R
\name{tunnelOpen}
\alias{tunnelOpen}
\title{Determine whether a tunnel opened by tunnelr is running}
\usage{
tunnelOpen(con)
}
\arguments{
\item{con}{An object of class "tunnel" returned by \code{tunnel()}.}
}
\value{
\code{TRUE} if the SSH tunnel is open, \code{FALSE}
  otherwise.
}
\description{
Determine whether a tunnel opened by tunnelr is running
}
\examples{
\dontrun{
sshTunnel = tunnel(remote_server, server_username,
                   remote_port, local_port)
tunnelOpen(sshTunnel) # returns TRUE
closeTunnel(sshTunnel)
tunnelOpen(sshTunnel) # returns FALSE
}
}

